assignment 5
Assignment Title: HTML, Basic OOP, and Superglobal Variables in PHP


 


Task 1: HTML Basics


Create an HTML form that allows users to input their name and email address. The form should have two fields: one for name and one for email. Use appropriate HTML tags to structure the form.


 


Task 2: Basic OOP in PHP


Create a PHP class called "Person" with the following properties and methods:


 


Properties:


 


Name


Email


Methods:


 


setName($name): sets the name property


setEmail($email): sets the email property


getName(): returns the name property


getEmail(): returns the email property


Create an instance of the Person class and set the name and email properties. Use the getName() and getEmail() methods to display the properties on the webpage.


 


Task 3: Superglobal Variables in PHP


Create a PHP script that retrieves the user's name and email from the HTML form in Task 1 using the $_POST superglobal variable. Create a new instance of the Person class and use the setName() and setEmail() methods to set the name and email properties based on the form data. Use the getName() and getEmail() methods to display the properties on the webpage.